requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
pandas==2.1.3
python-dateutil==2.8.2
streamlit==1.29.0
plotly==5.18.0
